refactor: abstract clicking on a list in separate hook
