+++ src/features/MovieList/hooks/useMovieList.ts
@@ -1,27 +1,20 @@
-import { MouseEvent, RefObject, useCallback, useRef } from 'react';
+import { RefObject, useCallback } from 'react';
 
 import LocomotiveScroll from 'locomotive-scroll';
-import { useNavigate } from 'react-router-dom';
 
 import {
   DEFAULT_MOVIES_PER_PAGE,
   DEFAULT_PAGE,
   MOVIES_PER_PAGE_PARAM,
   PAGE_PARAM,
-  SCROLL_TOP_DURATION,
 } from '../../../shared/const/const.ts';
 import useScrollTop from '../../../shared/hooks/useScrollTop.ts';
-import useTooltip from '../../../shared/hooks/useTooltip.ts';
 import useUrl from '../../../shared/hooks/useUrl.ts';
 import { MovieList } from '../../../shared/types/types.ts';
 import useSearch from '../../Search/hooks/useSearch.ts';
 
 function useMovieList(scroll: RefObject<LocomotiveScroll>) {
-  const listRef = useRef<HTMLUListElement>(null);
-  const navigate = useNavigate();
   const { movies, isLoading } = useSearch();
-
-  const { tooltipRef, hideTooltip, showTooltip } = useTooltip(scroll);
   const { readUrl, setUrl } = useUrl();
 
   const moviesPerPage = Number(
@@ -43,24 +36,9 @@ function useMovieList(scroll: RefObject<LocomotiveScroll>) {
     [setUrl],
   );
 
-  function handleClick(e: MouseEvent) {
-    const { target } = e;
-
-    if (target !== listRef.current) return;
-
-    navigate('/');
-    scroll.current?.scrollTo('top', { duration: SCROLL_TOP_DURATION });
-  }
-
   return {
     isLoading,
     moviesPerPage,
-    hideTooltip,
-    showTooltip,
-    tooltipRef,
-    listRef,
-    navigate,
-    handleClick,
     renderMovies,
     isNoMovies,
     handleMoviesPerPage,
